+++ src/components/content/hooks/use-toggle-archive.ts
@@ -1,5 +1,5 @@
 import { useCallback, useState } from "react";
-import { useCTW } from "@/components/core/providers/ctw-provider";
+import { useCTW } from "@/components/core/providers/use-ctw";
 import { toggleArchive } from "@/fhir/basic";
 import { FHIRModel } from "@/fhir/models/fhir-model";
 import { queryClient } from "@/utils/request";
@@ -32,16 +32,11 @@ export function useToggleArchive<T extends fhir4.Resource>(
 
   const handleToggleArchive = useCallback(async () => {
     setIsLoading(true);
-
-    toggleArchive(model, await getRequestContext()).then(
-      async () => {
-        setIsLoading(false);
-        await queryClient.invalidateQueries([queryToInvalidate]);
-      },
-      () => {
-        setIsLoading(false);
-      }
-    );
+    await toggleArchive(model, await getRequestContext());
+    await queryClient.invalidateQueries([queryToInvalidate]);
+    // Timeout here fixes bug where we would briefly flash
+    // the old dismiss/restore text.
+    setTimeout(() => setIsLoading(false), 0);
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [model]);