+++ src/components/content/timeline/patient-timeline.tsx
@@ -1,7 +1,7 @@
 import cx from "classnames";
 import { patientTimelineColumns } from "./patient-timeline-columns";
 import { usePatientEncounterDetailsDrawer } from "../timeline-2.0/helpers/modal-hooks";
-import { useCTW } from "@/components/core/providers/ctw-provider";
+import { useCTW } from "@/components/core/providers/use-ctw";
 import { Table } from "@/components/core/table/table";
 import { usePatientEncounters } from "@/fhir/encounters";
 import { orderBy } from "@/utils/nodash";