+++ src/components/core/providers/theme/context.ts
@@ -1,0 +1,12 @@
+import { createContext, createRef, RefObject } from "react";
+import { DefaultTheme, Theme } from "@/styles/tailwind.theme";
+
+export interface ThemeContextValue {
+  theme: Required<Theme>;
+  ctwThemeRef: RefObject<HTMLDivElement>;
+}
+
+export const ThemeContext = createContext<ThemeContextValue>({
+  theme: DefaultTheme,
+  ctwThemeRef: createRef<HTMLDivElement>(),
+});