+++ src/services/fqs/client.ts
@@ -1,7 +1,7 @@
 import { GraphQLClient } from "graphql-request";
 import { getZusApiBaseUrl } from "@/api/urls";
 import { CTWRequestContext } from "@/components/core/providers/ctw-context";
-import { Env } from "@/components/core/providers/ctw-provider";
+import { Env } from "@/components/core/providers/types";
 import { CTW_REQUEST_HEADER } from "@/utils/request";
 
 export interface GraphqlPageInfo {